This code has BLE working across all screens and prints the data for all sensors on the Connect Device Screen
This works by creating BLEUniversal file, which keeps the state and connection alive globally.

In App.tsx we are wrapping the rendered app in the BLEProvider (Creates one shared instance of BLEManager), which enusres that all screens have access to useBLE() (how your screens grab access to that shared manager + state)
